# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Controle de Hábitos API",
    "version": "1.0.0",
    "description": "API para registro, login, gerenciamento e histórico de hábitos do usuário."
  },
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário registrado" },
          "409": { "description": "Usuário já existe" }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido, retorna token" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/habits": {
      "post": {
        "summary": "Cadastrar hábito",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "habitName": { "type": "string" }
                },
                "required": ["habitName"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Hábito cadastrado" },
          "409": { "description": "Hábito já cadastrado" }
        }
      }
    },
    "/habits/{habitName}": {
      "delete": {
        "summary": "Deletar hábito",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "habitName",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }],
        "responses": {
          "200": { "description": "Hábito deletado" },
          "404": { "description": "Hábito não encontrado" }
        }
      }
    },
    "/habits/{habitName}/complete": {
      "post": {
        "summary": "Marcar hábito como concluído",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "habitName",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }],
        "responses": {
          "200": { "description": "Hábito marcado como concluído" },
          "400": { "description": "Hábito não encontrado ou já concluído" }
        }
      }
    },
    "/habits/history": {
      "get": {
        "summary": "Consultar histórico de hábitos concluídos",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de hábitos concluídos" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}